Share cmd and opts between `setup.ts` and `slice_helpers.ts`
